Refatora a lógica de experiência e classes, movendo dados para o novo sistema de subclasses. Ajusta a interface do usuário desabilitando classes de UI antigas e otimizando a exibição de proficiências. Melhora a legibilidade do código ao remover trechos obsoletos e simplificar callbacks.

+++ Common/GlobalClasses/RPGGlobalTile.cs
@@ -27,7 +27,7 @@ namespace Wolfgodrpg.Common.GlobalClasses
             // Se for uma planta colhível
             if (HarvestablePlantIDs.Contains(type))
             {
-                rpgPlayer.AddClassExperience("explorer", 5f); // XP por colher
+                rpgPlayer.AddClassExperience("explorer", (int)5f); // XP por colher
             }
             else // Se for qualquer outro bloco
             {
